cds-api/requirements.txt
fastapi
uvicorn
cdsapi
python-dotenv
xarray
netCDF4
cfgrib